# avoid creating a promise