this pr:

```
$ autocannon -c 100 -d 5 -p 10 localhost:3000
Running 5s test @ http://localhost:3000
100 connections with 10 pipelining factor

Stat         Avg     Stdev   Max
Latency (ms) 3.62    11.76   291.03
Req/Sec      27051.2 1275.73 28229
Bytes/Sec    4.42 MB 233 kB  4.63 MB

135k requests in 5s, 22.2 MB read
```

master:

```
$ autocannon -c 100 -d 5 -p 10 localhost:3000
Running 5s test @ http://localhost:3000
100 connections with 10 pipelining factor

Stat         Avg    Stdev   Max
Latency (ms) 3.78   11.99   250.12
Req/Sec      25912  1670.52 27637
Bytes/Sec    4.3 MB 277 kB  4.53 MB

130k requests in 5s, 21.2 MB read
```

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [x] commit message and code follows [Code of conduct](https://github.com/fastify/fastify/blob/master/CODE_OF_CONDUCT.md)