@allevo 2018-05-01
I don't understand the benchmarks. `Before` is 135k, `After` is 130k. What have I lost?