@StarpTech 2018-05-01
The same for https://github.com/fastify/fastify/blob/master/lib/reply.js#L295 ?