@mcollina 2018-05-02
@StarpTech that's not sitting in a major hot path, so I would not bother. But yes, you are right it might improve to have it there as well.